refactor(arch): split god objects into focused single-responsibility modules
Extracted from setupModel.ts (609 → 459 lines):
- src/features/game-setup/setupPresentation.ts: labels, rule descriptions,
  rules content per game type (SetupRulesContent type + 6 pure functions)

Extracted from App.tsx (555 → 347 lines):
- src/lib/analyticsInstance.ts: analytics port singleton
- src/app/useGameLifecycle.ts: all game lifecycle handlers (handleGameSelect,
  handleStartSetup, handle*Finish, handleRematch, handleReturnToGameSelection)

Extracted from MatchView.tsx (892 → 848 lines):
- src/features/x01/hooks/useMatchTimer.ts: elapsed timer + live clock
- src/features/x01/hooks/useMatchShortcuts.ts: shortcut state, draft sync,
  resize listener, handleShortcutDraftChange, resetShortcutDraft

Updated callers:
- views/SetupView.tsx: split import between setupModel + setupPresentation
- tests/unit/gameSetup/setupModel.test.ts: idem

Zero functional change. 120/120 unit tests passing.
